requirements.txt
asgiref==3.5.0
codegen==1.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-froala-editor==4.0.11
djangorestframework==3.13.1
djangorestframework-gis==0.3
env==0.1.0
filelock==3.4.2
Markdown==3.3.6
patterns==0.3
Pillow==9.0.1
platformdirs==2.4.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.12.0
